backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
pydantic
